Seven day: Writing some models with consume, converting images to bit format, and making improvements to API Edit methods.

+++ Final Project API/API-FinalProject/Service/Services/AboutBannerImgService.cs
@@ -54,9 +54,12 @@ namespace Service.Services
             if (existImg == null) throw new KeyNotFoundException($"Image with ID {id} not found.");
             if (model.Img != null)
             {
-                string oldFileName = Path.GetFileName(existImg.Img);
-                _fileService.Delete(oldFileName, "aboutbannerimgs");
 
+                if (!string.IsNullOrEmpty(existImg.Img))
+                {
+                    string oldFileName = Path.GetFileName(existImg.Img);
+                    _fileService.Delete(oldFileName, "aboutbannerimgs");
+                }
                 string newImageUrl = await _fileService.UploadFileAsync(model.Img, "aboutbannerimgs");
                 existImg.Img = newImageUrl;
             }
